requirements.txt
hltv-async-api>=0.9.0
requests>=2.31
PyYAML>=6.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3
